#156614968  Update-room-info  (#11)
* [Feature #156614968]Add update room mutation

* [Feature #156614968 ]Add tests

* [Feature #156461691] Change os.environ.get to os.getenv
Squashing  structure commits

[Feature #156461691] Change manage.py file

[Feature #156461691] Change database configurations to database module

[Feature #156461691] Implement schema for room and equipment modules

[Feature #156461691] Implement location, block, floor, room and equipment models

[Feature #156461691] Add .gitignore file

[Feature #156461691] Implement test of room creation

[Feature #156461691] Add fixtures file

[Feature #156461691] Change mock mutation data to fixtures file

[Feature #156461691] Modularise models in different packages

[Feature #156461691] Add fixtures helpers and utilities packages

[Feature #156461691] Add test_room package for room tests

[Feature #156461691] Refactor to import database from helpers

[Feature #156461691] Refactor to include helpers and models

[Feature #156461691] Refactor to have general schema

[Feature #156461691] Delete files

Update database.py

[Feature #156461691] Modularise models in different packages

[Feature #156461691] Delete files

[Feature #156461691] Refactor files

[Feature #156461691] Change FLASK_CONFIG to APP_SETTINGS

[Feature #156461691] Change FLASK_CONFIG to APP_SETTINGS

* [Feature #156614968]Finalize squashed changes

* [Fix #156614968]Clean up code

* [Feature #156614968]Add validation for empty string input

This commit caters for validation of empty strings incase you are updating the fields. The update query has also been updated to query by room id hence the tests have been changed to reflect this change. A new function has been placed in the utilities.py to cater for input validation

* Add untracked  files after merge

* [Fix #156614968]Add /__pycache__/ to .gitignore

* [Feature #156614968]Add update room mutation

* [Feature #156614968 ]Add tests

* [Feature #156614968]Finalize squashed changes

* [Fix #156614968]Clean up code

* [Feature #156614968]Add validation for empty string input

This commit caters for validation of empty strings incase you are updating the fields. The update query has also been updated to query by room id hence the tests have been changed to reflect this change. A new function has been placed in the utilities.py to cater for input validation

* [Fix #156614968]Add /__pycache__/ to .gitignore

* [Fix #156614968]Clean up __pycache__files

* [Fix #156614968] Rename equipmet to resource

* [Feature #156614968]Update mutation

 This commit updates the Update Room feature so as to get rid of relay and node and also to update using roomId

* [Fix #156614968]Fix import error

* [Fix #156614968]Fix merge conflict

* [Fix #156614968]Fix failing tests

* [Fix #156614968] Add update entity method

This commit  refactors the update mutation so that we do not have similar code and also introduces the update_entity_method in the utilities that can be used by all update mutations for various entities. Tests have been changed to fix fails.

* [Fix #156614968] Add update entity method

This commit  refactors the update mutation so that we do not have similar code and also introduces the update_entity_method in the utilities that can be used by all update mutations for various entities. Tests have been changed to fix fails.

* [Fix #156614968]Resolve to use assertEquals

 To improve code readability , this commit changes the test mode of assertion from assert to assertEquals

* [Feature #156614968]Add test

This commit adds testing for an empty  field for example and empty string when updating a name.