#157385654 View room schedule (#30)
* Add untracked  files after merge

* [Fix] Pull Updated changes to develop

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Fix #157385654]Remove _pycache_ files

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add latest changes from develop

* [fix #157385654]Fix flake8 issues

* [Fix #157385654] Fix failing test.

* [Feature #157385654]Add tests for events

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add latest changes from develop

* [Fix #157385654] Fix failing test.

* [Feature #157385654]Add tests for events

* [Fix #157385654]Run migrations with calendar_id field

* [Fix #157385654]Pull remote changes

* [Fix #157385654]Fix failing test fixtures

* [Fix #157385654]Fix flake8 errors

* [Fix #157385654]Fix migration issues

* [Fix #157385654] Add descriptive error

This commit enables raising of a more descriptive error when a calandarId given on the query is not assigned to any calendar on  ocnverge

* [Fix]Add google api credentials to .env example

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654]Add latest changes from develop

* [Fix #157385654] Fix failing test.

* [Feature #157385654]Add tests for events

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add tests

Add tests for the RoomSchedule query that contains google calendar intergration

* [Feature #157385654] Add getRoomSchedule query

* [Feature #157385654]Add view room schedule functionality

This commit  enables a user to view room schedule from the google calendar api. One is able to specify the calendar_id and the number of days they need and will then receive information containing the event start time and summary of the event

* [Feature #157385654]Add latest changes from develop

* [Fix #157385654] Fix failing test.

* [Feature #157385654]Add tests for events

* [Fix #157385654]Run migrations with calendar_id field

* [Fix #157385654]Pull remote changes

* [Fix #157385654]Fix failing test fixtures

* [Fix #157385654]Fix flake8 errors

* [Fix #157385654]Fix migration issues

* [Fix #157385654] Add descriptive error

This commit enables raising of a more descriptive error when a calandarId given on the query is not assigned to any calendar on  ocnverge

* [Fix]Add google api credentials to .env example

* [fix] Add new changes from develop

* [fix] Add new changes from develop

* [Fix]Fix failing test